+++ extensions/bitwarden/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Bitwarden Changelog
 
+## [Fix] - 2024-10-04
+
+- Simplify number and special effects on min fields to prevent re-render loop
+
 ## [Fix and New Password Generation Options] - 2024-10-02
 
 - Added Minimum numbers and Minimum special password options